+++ .vscode/launch.json
@@ -15,7 +15,7 @@
                 "999999",
                 "-r",
                 "ts-node/register",
-                "${workspaceFolder}/src/test/**/*.ts"
+                "${workspaceFolder}/src/test/**/*.ts", "-g", "getDocumentSymbols"
             ],
             "internalConsoleOptions": "openOnSessionStart"
         }
