+++ src/test/externalConsumption.test.ts
