+++ src/test/inspection/activeNode.test.ts
