+++ src/test/inspection/currentInvokeExpression.test.ts
