+++ src/test/inspection/scope.test.ts
