+++ src/test/inspection/type.test.ts
