+++ src/test/providers/simpleLibrarySymbolProvider.test.ts
