+++ src/test/testUtils/index.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export * from "./abridgedUtils";
-export * from "./analysisUtils";
-export * from "./assertEqualUtils";
-export * from "./inspectionUtils";
+export * from "./abridgedTestUtils";
+export * from "./analysisTestUtils";
+export * from "./assertEqualTestUtils";
+export * from "./inspectionTestUtils";
 export * from "./testUtils";
